requirements.txt
anthropic>=0.50.0
tavily-python>=0.5.0
streamlit>=1.35.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
